Rename unexpected.simon.js to unexpected-simon.js
