+++ lib/unexpected-sinon.js
