+++ package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "test": "./node_modules/.bin/mocha"
   },
-  "main": "./lib/unexpected.sinon.js",
+  "main": "./lib/unexpected-sinon.js",
   "dependencies": {
     "sinon": "*",
     "unexpected": "*"
