+++ test/common.js
@@ -1,4 +1,4 @@
 sinon = require('sinon');
 expect = process.env.COVERAGE ?
-    require('../lib-cov/unexpected.sinon') :
-    require('../lib/unexpected.sinon');
+    require('../lib-cov/unexpected-sinon') :
+    require('../lib/unexpected-sinon');
