+++ test/tests.html
@@ -23,8 +23,8 @@
     <script>
       expect = weknowhow.expect;
     </script>
-    <script src="../lib/unexpected.sinon.js"></script>
-    <script src="unexpected.sinon.spec.js"></script>
+    <script src="../lib/unexpected-sinon.js"></script>
+    <script src="unexpected-sinon.spec.js"></script>
     <script>
       mocha.run();
     </script>
